Remove redundant `set nocompatible` from vimrc
When the compatible option is off, Vim settings are used rather than Vi
settings. Adding `set nocompatible` is doubly redundant because:
1. The presence of a vimrc file causes the compatible option to be off
   by default.
2. Sourcing $VIMRUNTIME/defaults.vim also sets nocompatbile, with a
   with a guard clause that checks for `&compatible`.

The one case in which compatible might still be turned on is if I were
to start vim with the `-u` option but not the `-N` option. It does not
seem worth protecting against this case.