Gitconfig: Set git pull default to --ff-only
In Git 2.27, a new warning was added that discourages using `git pull`
without specifying how to reconcile divergent branches. This commit
specifies to use the `--ff-only` option which completes the pull only if
it can be resolved as a fast-forward merge. When a fast-forward merge is
not possible, creating a new merge commit or rebasing will be required.