vimrc: Enable italics
This commit sets the `t_ZH` and `t_ZR` terminal output codes which
define italics mode and italics end, respectively, so that italics are
properly supported. Note that the `` characters are literal escape
characters which must be entered using `<Ctrl-v><Esc>`.

Additionally, I had to edit `.vim/colors/solarized.vim` as recommended
in altercation/vim-colors-solarized/#120 so that solarized recognized
that iTerm2 supports italics.